Add code for the “Incorporating Accessibility Into Your Development Process” bit
This takes all of the “problem” code from the other accessibility bit and turns it into React components.

My goal with this commit is to move the code into the new bit’s folder and use it for testing different accessibility tools. I’ll probably move this code into an exercise, or this will be part of one mega exercise. I will figure it out. 😃