Add jest-axe to the “Incorporating Accessibility” demo
The `jest-axe` package is awesome because it can test your individual components (or your whole app) for accessibility violations.

This runs the tests against the whole app so I can easily compare tools. If we use this in an exercise, I will encourage folks to add this to their individual component tests.